Ignore statsbuckets for text/varchar/bpchar/char column data types.
We added a utility function `FNDVBasedCardEstimationDatum()` that checks
the column's data type and returns a BOOL that we use to do NDV Based
cardinality estimation for the above column data types.

We made `PhistLessThanOrLessThanEqual()` and `PhistGreaterThanOrGreaterThanEqual()`
use the check to determine if it should use NDVRemain and FreqRemain for
cardinality estimation.