Collation support for CScalarCast & CMDType*
- Add collation field to CScalarCast.

- Add default collation oid to CMDType*. For Int2, Int4, Int8, Oid and Bool
almost all of the metadata information is hard-coded in their respective
header files. Hence for these types, we hardcode the default collation as well.
For generic types however, we need to get the default collation oid from
Relcache Translator and store it in the cache. Corresponding changes have been
made in the Relcache Translator.

Signed-off-by: Dhanashree Kashid <dkashid@pivotal.io>