Add collation support to CScalarIf
CaseExpr in postgres is translated into two different expressions in
Orca.
When a test expression is present:
e.g. CASE b WHEN 'abc' THEN ...
CaseExpr is translated into a CScalarSwitch
When a test expression is not present and there is a bool expression:
e.g. CASE WHEN b = 'abc' THEN ...
CaseExpr is translated into a CScalarIf

This commit adds collation to CDXLScalarIf and CScalarIf

Signed-off-by: Jacob Champion <pchampion@pivotal.io>