Add collation to CScalarCaseTest and CScalarSwitch
To support the second form of CaseExpr (with a CaseTestExpr)
add collation to CScalarCaseTest and CScalarSwitch as well as
CDXLScalarCaseTest and CDXLScalarSwitch to support Postgres CaseExpr
with testexpr Also, parse and serialize collation. Added collation to
CParseHandlerScalarCaseTest and did a minor refactor, removing the
return type, as it does not seem needed.

Signed-off-by: Omer Arap <oarap@pivotal.io>