Hungarian notation removal: GPOS and Naucrates libraries
This is the first commit of the initiative to remove Hungarian notation
and replace it with a new convention -- a custom guidelines based on
common C++ conventions. The style guide for this convention can be found
in StyleGuide.md.

NB: This does not cover all ORCA source code. The other libraries are
planned to be converted in the future.

Also bump ORCA version to 2.69.0

 Co-authored-by: Melanie Plageman <mplageman@pivotal.io>
 Co-authored-by: Ekta Khanna <ekhanna@pivotal.io>
 Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>
 Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
 Co-authored-by: Sambitesh Dash <sdash@pivotal.io>
 Co-authored-by: Dhanashree Kashid <dkashid@pivotal.io>
 Co-authored-by: Omer Arap <oarap@pivotal.io>