Account for replicated distribution when setting Input segments for Gather motion
If the tree underneath of Physical Gather Motion has a distribution spec
as "Replicated" then, the number of input segments for Gather are set to
1 by `FDuplicateHazardMotion()`. However `CheckValidity()` did not allow
Gather Motion with only one Input Segments. Now with replicated tables,
there will be scenarios where Gather needs to happen with a single
segment as input.