Add flag to disable motion between Global and Local
CPhysicalAgg checks the distribution of a local aggregate and prohibits
a plan if the local agg distribution is the same as the one requested by
its parent (i.e. local and global have same distribution).

This is the right approach for SplitAgg Xform since the grouping columns
for the global and local are the same. For EagerAgg, however, the local
aggregate adds additional grouping columns, hence making it valuable to
have the aggregate split even with the same distribution.

This is achieved by adding a flag within CPhysicalAgg (and its
subclasses). The flag is set (i.e. = true) by default, indicating that
the distribution check is enforced in all cases except EagerAgg. The
flag is unset during the Logical -> Physical transformation when the
operator was created by the EagerAgg Xform.