Include binary coercible casts for predicate inference
This comments augments ORCA's predicate inference system, by also
considering binary coercible casted scalar idents & constants, both on
the lhs and rhs of an expression. The logic strips the cast and uses
the underlying CColRef in equiv classes. It also passes the CColRef to
the constraint framework. This way the logic in in the constraint
framework remains unchanged.

Note, however, that when the inferred constraint is converted back to an
expression, appropriate casts (binary coercible or otherwise) must be
added back. This is handled in CUtils::PexprScalarCmp(). Once we
retrieve the appropriate operator to use for construct the scalar cmp,
we add necessary casts to both the left and right expressions,
according to the left & right oids required by the scalar cmp operator.