Apply correct casts to match operand mdids for ScCmp
GetScCmpMdIdConsiderCasts() returns the mdid of the operator to use to
construct the scalar cmp op, but doesn't modify the expressions. This
commit adds a new method GetScCmpMdIdApplyCasts() that determines the
correct cast(s), if any, to add on the left and right expressions. It
does this by comparing the mdids of the expressions and the input mdids
expected by the operator.

For example, when asked for =(varchar, varchar), if the op returned by
GetScCmpMdIdConsiderCasts() is =(text, text), both the lhs and rhs must
be cast to text.