Fix for cardinality estimation of left outer joins
The cardinality of a left outer join is estimated as the cardinality of
an inner join plus the cardinality of a left anti-semijoin (LASJ).

The cardinality of the LASJ is computed by subtracting the inner child
histogram from the outer child, leaving only buckets of the outer table
that have no counterparts in the inner table. This is computed by method
CHistogram::MakeLASJHistogram().

There was a bug in this method where we stopped too early in some
cases, leading to an over-estimation of the rows in the LASJ.

We also fixed a couple of small issues in CBucket::CompareLowerBounds(),
a typo and two cases where we compared the boundary values but not
the included/excluded flags.

Several mdp files needed changes, all of these now return lower
cardinality estimates for left outer joins.

Note that we may still overestimate some outer join cardinalities,
since we use an NDV-based estimate for the inner join and this
estimate sometimes exceeds the number of rows indicated by the
intersection of inner and outer histogram. When that is the case,
then adding the difference of outer - inner histograms causes
us to double-count the rows in the LASJ. We can consider changing
that in a future fix, if it becomes an issue with plan quality.

Co-authored-by: Bhuvnesh Chaudhary <bchaudhary@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>