Don't extract all bindings below scalar operators
Consider the below input query tree:
```
+--CLogicalInnerJoin
   |--CLogicalGet "t1"
   |--CLogicalSelect
   |  |--CLogicalGet "t3"
   |  +--CScalarSubqueryAny(=)
   |     |--CLogicalUnion
   |     |  |--CLogicalGet "t4"
   |     |  +--CLogicalGet "t5"
   |     +--CScalarIdent "b"
   +--CScalarCmp (=)
      |--CScalarIdent "b" (1)
      +--CScalarIdent "b" (10)
```
For the Operator CLogicalSelect, there are 2 different alternatives for
scalar condition which are generated by CXformSimplifySelectWithSubquery:
    1. CScalarSubqueryAny(=)["b" (19)]
    2. CScalarCmp (>)

For the Xform, CXformInnerJoinWithInnerSelect2IndexGetApply (and other equivalent xforms),
the pattern is as below:
```
+--TJoin
   |--CPatternLeaf
   |--CLogicalSelect
   |  |--TGet
   |  +--CPatternTree
   +--CPatternTree
```
If we look at the pattern for the scalar condition of CLogicalSelect, it's
defined as a CPatternTree. For the test query, the inner child of the join
(i.e CLogicalSelect), can have 2 different representation for the scalar
condition as denoted below by Alternative 1 and 2.  Anything below the
Scalar operators (CScalarSubqueryAny, CScalarCmp) is irrelevant in
determining the indexes which can be used for CLogicalGet "t3". Also, there
could be several bindings below the scalar operator, for instance, a Union
Operation can also be performed using Agg (Global) on Union ALL, and the
Global Agg could further have Global and Local alternatives.  We should
extract the Scalar Group Expression only once per alternative. In the case
of the test query used, there are 2 alternatives, so the test expects that
there should be 2 bindings for the pattern defined by
CXformInnerJoinWithInnerSelect2IndexGetApply.

Alternative: 1
```
+--CLogicalSelect
   |--CLogicalGet "t3"
   +--CScalarSubqueryAny(=)
      |--CLogicalUnion
      |  |--CLogicalGet "t4"
      |  +--CLogicalGet "t5"
      +--CScalarIdent
```

Alternative: 2 (CScalarSubqueryAny transformed to CScalarCmp > CScalarSubquery)
```
+--CLogicalSelect
   |--CLogicalGet "t3"
   +--CScalarCmp (>)
      |--CScalarSubquery
      ...... (Subtree - removed as its not relevant)
```
Bump ORCA to v3.31.0
Signed-off-by: Ashuka Xue <axue@pivotal.io>