Reduce colrefs by skipping unused ones
Currently we retain all the colrefs of a table even though we only reference a
few of them in the original query. This patch marks the colrefs as used if
either its a system column or if its referenced in the query and unused
otherwise. When we derive the relational properties, we exclude the unused
columns from the output columns list. For a SELECT COUNT(*) query on a 1600
column table which has a 10 way join, we see an improvement of approximately
40%.

Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>