Fix PdsRequired for local CPhysicalAgg with DQAs
When we had a local aggregate with a DQA (e.g. select count(distinct b)
... group by a), we generated a required distribution on the DQA column
(b in this example). We now also consider an alternative distribution
spec with both the groupby and the DQA columns (a, b in the example).

Note: We have some MDP tests for multiple DQAs which are not supported
in ORCA - when run through PSQL the same query will fall back. We
don't generate good plans for those and we should probably delete
those MDPs (MDQA*.mdp).

Co-authored-by: Sambitesh Dash <sdash@pivotal.io>