Introduce aggregated memory allocation, using dlmalloc
ORCA does a lot of memory allocations, and each allocation is fairly expensive,
recording additional information and doing checks (e.g. OOM).

We would like to use the GPDB memory allocators by default (change the default
for optimizer_use_gpdb_allocators to on), to prevent ORCA from crashing when
we run out of memory (OOM). However, with the current way of doing things, doing
so would add around 10 % performance overhead to ORCA. This is not because the
GPDB memory allocators are inefficient, but because we are calling them very
many times.

As a solution, we are adding an "aggregation layer" to the ORCA CMemoryPoolTracker
to speed up memory allocations. This will allow us to use GPDB memory allocators
while at the same time speeding up ORCA overall.

Basically, this aggregation layer manages memory and combines many (usually small)
memory requests into larger requests that are then satisfied by the GPDB memory
allocators. The code to do this comes from open source, a memory allocation algorithm
called dlmalloc, developed by Doug Lea that is released under Creative Commons CC0
Public Domain Dedication. A similar version of this code is also used by the GNU
runtime library.

See http://gee.cs.oswego.edu/dl/html/malloc.html and
http://gee.cs.oswego.edu/pub/misc/malloc-2.8.6.c

On our development machine (MacOS), we saw 15 % to 40 % improvement in optimization
time and about a 50 % reduction in the allocated memory. This is using a retail build.

We made this change in the CMemoryPoolTracker class, since it is the one that is used
almost everywhere as the top-most memory pool object. Other classes like CMemoryPoolAlloc
are used in lower layers to do the actual allocation of memory.

What did we change in dlmalloc?

- We eliminated global variables in dlmalloc and put a struct describing the state
  of the dlmalloc system into each CMemoryPoolTracker object. This struct is called
  malloc_state.
- This means that each CMemoryPoolTracker object aggregates memory independently.
- Since ORCA does not specify a memory pool when deleting memory, we need to store
  a pointer back to the malloc_state in every allocated chunk of memory.
- We tried to simplify the dlmalloc code by deleting many unused features.
- The dlmalloc method requests memory in larger chunks from the lower layers
  (CMemoryPoolAlloc and ultimately the GPDB memory allocators). These chunks are
  called "segments". To avoid waste and having too many segments, the segment size
  increases exponentially, starting at 32 KB and increasing in factors of 8, up to
  64 MB.

The code is written to allow a mix of aggregated and non-aggregated requests, even
in the same CMemoryPoolTracker. When deallocating, we recognize which type of allocation
we used by a marker (CMemoryPool::AllocHeader::m_zero_marker).

For array allocations, we store the number of array elements in the allocated chunk,
to be able to call the correct destructors when deallocating the memory.

We also made the parameters that pass file name and line numbers to the memory allocation
routines conditional, they are now only passed when GPOS_DEBUG is defined. Finding memory
leaks will now be slightly different: The initial step will be the same, when running
an MDP file in gporca_test, a list of memory leaks will be generated, but the file name
and line numbers point to the aggregation layer in CMemoryPoolTracker. To find the actual
objects that are leaked, run gporca_test with the following option to turn off aggregated
allocation: -T 105. In psql, use this guc:

set optimizer_enable_aggregated_memory_allocation = off;

This feature will be enabled by default in 6X and Master. It will be off in 5X and there
is no guc to turn it on in 5X.

When parsing DXL, a memory pool is created before the trace flags are set. In that case,
we need to update the memory pool to disable aggregated memory, once the traceflag is
set.

Misc. changes:

- Avoid using an AutoMemoryPool in one instance, use the default memory pool of CTask::Self()
  instead, this should avoid allocating unnecessary extra segments.

- Avoid exception when loading search strategy.

Co-authored-by: Ashuka Xue <axue@pivotal.io>
Co-authored-by: Chris Hajas <chajas@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>