Partial revert of PR 530
The issue solved by PR 530 stemmed not from costing, but from an index
path that was not expected to be followed. With the past commit which
combines the path for col op col and col op const. The alternative that
needed to be avoided does not get generated. As such, PR 530 can be
reverted so that the bitmap costing is more accurate.

Co-Authored-By: Ashuka Xue <axue@pivotal.io>
Co-Authored-By: Hans Zeller <hzeller@pivotal.io>