Debugging code to count events, quantities, CPU time
This code makes it easy to count and measure various things in ORCA,
assuming you are using a debug build and are willing to modify the code.

The commit adds a new class CDebugCounter, which is enabled only in debug
builds (although that can be changed temporarily to measure retail builds,
by defining GPOS_DEBUG_COUNTERS).

You instrument the code by using the following macros to count events,
add up quantities (integers and floating point values) and to count CPU
time:

  GPOS_DEBUG_COUNTER_BUMP(counter_name)
  GPOS_DEBUG_COUNTER_ADD(counter_name, delta)
  GPOS_DEBUG_COUNTER_ADD_DOUBLE(counter_name, delta)
  GPOS_DEBUG_COUNTER_START_CPU_TIME(counter_name)
  GPOS_DEBUG_COUNTER_STOP_CPU_TIME(counter_name)

This commit adds code to ORCA to collect this information and write it
to the event log at the end of each statement compilation (again, this
is only enabled for debug builds unless changed).

You can have different counters, each one with a name chosen by you.
However, a given counter name can be used only with one of the measurement
types (count, sum of integers, sum of doubles, CPU time).

When you run multiple queries, ORCA will report the counter values separately
for each query. Any totals would have to be done by you with an SQL query
or a spreadsheet.

For convenience, you can name queries in a file, like so:

   select 'query name: q1';
   explain select * from foo;

In other words, a simple constant get with a string that starts with
"query name: " will be recognized by ORCA to indicate the name of the
next query. The counters will not be reported fir this first select,
only for the second query. Query names are useful when you have a test
with very many queries in it, e.g. the TPC-DS queries.

Finally, the commit contains a Python program to retrieve the recorded
counters from the log. The program can either print a series of comma-
separated values (CSV) to stdout or it can insert the result into an
SQL table name debug_counters in a database you can choose.

Run this for more info:

.../gporca/scripts/get_debug_event_counters.py --help