Better handling of null-rejecting predicates for ANDs and ORs
The methods that recognize null-rejecting predicates didn't handle
the case where an operand could be false or null, but not true.

We introduced a new Enum EberNotTrue to handle this case.

We also renamed the EberUnknown enum to EberAny, to avoid confusion
with the concept of "unknown" in the SQL standard, which means
something very similar to "null".

Co-authored-by: Abhijit Subramanya <asubramanya@pivotal.io>
Co-authored-by: Hans Zeller <hzeller@pivotal.io>