Disable event counter code for Linux
Temporarily disabling this for Linux, as it probably is causing
the CentOS6 and CentOS 7 build to fail in the pipeline.

We might want to fix those in a different way in the future, for example
by making sure we use the same C++ library for both cmake and ctest.
Today, the concourse/build_and_test.py script potentially allows different
libraries to be used.