Consider Bitmap OR of BTree indexes for ORs of index predicates
When we have predicates that are too complicated for BTree index scans,
then also consider bitmap index scans. This is implemented by passing
a boolean argument, indicating that we are processing a complex predicate
and by adding an additional parameter to allow btree indexes in addition
to the primary index type, bitmap.

The change also required updating some older MDP files with the return
type of the index, since this is now required for BTree indexes as well
as for bitmap indexes (that has been the case since commit 976cba9b0).