Don't allow btree indexes in bitmap when breaking up AND predicates
Most AND predicates are handled by the btree code path ok, so there
is not need to consider them in the bitmap code path, since we could
cost the bitmap scans cheaper than the btree index scans.

Note that there are three main cases of ANDed predicates and btree indexes:

- Multiple predicates for the same index, this is handled by the
  btree code path
- Predicates for different indexes, ORCA will pick only one index per
  alternative, so it will generate the cheapest plan that uses one index,
  while a bitmap scan would be able to use multiple indexes
- Predicates for one index plus non-index (residual) predicates, this
  is handled by the btree code path

In summary, we have an "opportunity to improve" for ANDed predicates on
multiple btree indexes. We won't fall back for such predicates and we
won't generate a bitmap scan at the moment (as planner does). This is
not a regression.