Fixes and changes for OsPrint and DbgPrint methods
In an earlier PR I broke the DbgPrint method on CColRefSet.
This includes the fix - removing an extra declaration of DbgPrint.

Adding a template-based way to print the contents of CDynamicPtrArray
when the data type contained in the array contains an OsPrint method.

Another refactor: Avoid use of array of fn ptrs in CDecorrelator::FProcessOperator

Co-authored-by: Jesse Zhang <jessezhang@vmware.com>